# backend/src/main/resources/openapi.yaml
openapi: 3.1.1
info:
  title: ちいぴよ API
  description: 子供の写真・動画管理及び育児記録管理アプリケーションのAPI定義
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - BearerAuth: []
...
paths:
...
  /sharing-groups:
    get:
      tags:
        - sharingGroupManagement
      summary: 共有グループ一覧を取得
      operationId: getSharingGroups
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      responses:
        '200':
          description: 共有グループ一覧
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SharingGroupResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    post:
      tags:
        - sharingGroupManagement
      summary: 共有グループを新規作成
      operationId: createSharingGroup
      x-codegen-request-body-name: SharingGroupData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        description: 作成する共有グループ情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/SharingGroupRequestDto'
        required: true
      responses:
        '201':
          description: 共有グループ作成成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /sharing-groups/{id}:
    get:
      tags:
        - sharingGroupManagement
      summary: 共有グループをID指定で取得
      operationId: getSharingGroup
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 共有グループ情報
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/SharingGroupResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    patch:
      tags:
        - sharingGroupManagement
      summary: 共有グループを編集
      operationId: updateSharingGroup
      x-codegen-request-body-name: SharingGroupUpdateData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        description: 編集する共有グループ情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/SharingGroupUpdateRequestDto'
        required: true
      responses:
        '200':
          description: 更新された共有グループ情報
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/SharingGroupResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - sharingGroupManagement
      summary: 共有グループを削除
      operationId: deleteSharingGroup
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: 共有グループの削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
...
  /media/{id}:
    get:
      tags:
        - mediaManagement
      summary: メディアをID指定で1件取得
      operationId: getMedia
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: メディア情報
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/MediaResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    patch:
      tags:
        - mediaManagement
      summary: メディア情報を更新
      operationId: updateMedia
      x-codegen-request-body-name: MediaUpdateData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        description: 更新するメディア情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/MediaUpdateRequestDto'
        required: true
      responses:
        '204':
          description: メディアの更新に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - mediaManagement
      summary: メディアを削除
      description: ゴミ箱に移動し30日間保持
      operationId: deleteMedia
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: メディアの削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
...
  /media/{mediaId}/tags:
    put:
      tags:
        - tagManagement
      summary: メディアのタグを一括更新
      operationId: updateMediaTags
      x-codegen-request-body-name: MediaTagsData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/MediaId'
      requestBody:
        description: 紐付けるタグIDの一覧
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/MediaTagsUpdateRequestDto'
        required: true
      responses:
        '200':
          description: 更新後のタグ一覧
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /tags:
    get:
      tags:
        - tagManagement
      summary: タグ一覧を取得
      operationId: getTags
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      responses:
        '200':
          description: タグ一覧
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    post:
      tags:
        - tagManagement
      summary: タグを作成
      operationId: createTag
      x-codegen-request-body-name: TagData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        description: 作成するタグ情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/TagRequestDto'
        required: true
      responses:
        '201':
          description: 作成されたタグ
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/TagResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /tags/{id}:
    put:
      tags:
        - tagManagement
      summary: タグを編集
      operationId: updateTag
      x-codegen-request-body-name: TagUpdateData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        description: 編集するタグ情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/TagRequestDto'
        required: true
      responses:
        '204':
          description: タグの更新に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - tagManagement
      summary: タグを削除
      operationId: deleteTag
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: タグの削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognitoで取得したJWTトークンを指定
  parameters:
    XRequestedWith:
      name: X-Requested-With
      in: header
      required: true
      schema:
        type: string
        default: XMLHttpRequest
    ResourceId:
      name: id
      in: path
      description: リソースの一意な識別子
      required: true
      schema:
        type: integer
        format: int64
...
    MediaId:
      name: mediaId
      in: path
      description: メディアの一意な識別子
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ErrorResponseDto:
      type: object
      description: エラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: BAD_REQUEST
        message:
          type: string
          description: エラーメッセージ
          example: リクエストが不正です
...
    SharingGroupMemberResponseDto:
      type: object
      description: 共有グループメンバー情報
      required:
        - id
        - userId
        - displayName
        - presignedIconUrl
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: メンバーID
          example: 1
        userId:
          type: integer
          format: int64
          description: ユーザーID
          example: 2
        displayName:
          type: string
          description: ユーザー表示名
          example: ママ
        presignedIconUrl:
          type:
            - string
            - 'null'
          description: プロフィール画像のS3署名付きURL
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
    SharingGroupResponseDto:
      type: object
      description: 共有グループ情報
      required:
        - id
        - name
        - members
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: グループID
          example: 1
        name:
          type: string
          description: グループ名
          example: 家族全員
        members:
          type: array
          description: グループメンバー一覧
          items:
            $ref: '#/components/schemas/SharingGroupMemberResponseDto'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: '2026-01-01T00:00:00Z'
    SharingGroupRequestDto:
      type: object
      description: 共有グループ作成リクエスト
      required:
        - name
      properties:
        name:
          type: string
          description: グループ名
          minLength: 1
          maxLength: 100
          example: 家族全員
        userIds:
          type: array
          description: グループメンバーID一覧
          items:
            type: integer
            format: int64
    SharingGroupUpdateRequestDto:
      type: object
      description: 共有グループ更新リクエスト
      properties:
        name:
          type: string
          description: グループ名
          minLength: 1
          maxLength: 100
          example: 家族全員
        userIds:
          type: array
          description: 新しい共有グループメンバーID一覧
          items:
            type: integer
            format: int64
    TagResponseDto:
      type: object
      description: タグ情報
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: タグID
          example: 1
        name:
          type: string
          description: タグ名
          example: お出かけ
        mediaCount:
          type: integer
          format: int64
          description: タグが紐づいているメディアの数
          example: 10
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
    MediaNavigationResponseDto:
      type: object
      description: 前後のメディア情報
      required:
        - id
        - mediaType
      properties:
        id:
          type: integer
          format: int64
          description: メディアID
          example: 1
        thumbnailPresignedUrl:
          type:
            - string
            - 'null'
          format: uri
          description: サムネイルのS3アップロード用の署名付きURL
        mediaType:
          type: string
          description: メディア種別
          enum:
            - PHOTO
            - VIDEO
          example: PHOTO
    MediaResponseDto:
      type: object
      description: メディア情報
      required:
        - id
        - uploadedBy
        - mediaType
        - originalFilename
        - contentType
        - fileSize
        - width
        - height
        - takenAt
        - albumId
        - sharingGroupId
        - uploadStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: メディアID
          example: 1
        uploadedBy:
          type: integer
          format: int64
          description: アップロードしたユーザーID
          example: 1
        mediaType:
          type: string
          description: メディア種別
          enum:
            - PHOTO
            - VIDEO
          example: PHOTO
        originalFilename:
          type: string
          description: ファイル名
          example: IMG_20260101.jpg
        contentType:
          type: string
          description: MIMEタイプ
          example: image/jpeg
        fileSize:
          type: integer
          format: int64
          description: ファイルサイズ（バイト）
          example: 2048000
        width:
          type:
            - integer
            - 'null'
          description: 幅（ピクセル）
          example: 1920
        height:
          type:
            - integer
            - 'null'
          description: 高さ（ピクセル）
          example: 1080
        presignedUrl:
          type: string
          format: uri
          description: S3署名付きURL
        thumbnailPresignedUrl:
          type:
            - string
            - 'null'
          format: uri
          description: サムネイルのS3署名付きURL
        takenAt:
          type:
            - string
            - 'null'
          format: date
          description: 撮影日
          example: '2026-01-01'
        albumId:
          type:
            - integer
            - 'null'
          format: int64
          description: 所属アルバムID
          example: 1
        sharingGroupId:
          type: integer
          format: int64
          description: 共有範囲グループID
          example: 1
        uploadStatus:
          type: string
          description: アップロード状況
          enum:
            - PROCESSING
            - COMPLETED
            - FAILED
          example: COMPLETED
        addFavoriteUserIds:
          type: array
          description: お気に入りに追加したユーザーID一覧
          items:
            type: integer
            format: int64
          example:
            - 1
            - 2
            - 3
        tags:
          type: array
          description: 付与されているタグ一覧
          items:
            $ref: '#/components/schemas/TagResponseDto'
        isFavorite:
          type: boolean
          description: お気に入り済みフラグ
          example: false
        commentCount:
          type: integer
          format: int64
          description: コメント数
          example: 1
        nextMedia:
          type:
            - object
            - 'null'
          description: 1つ後のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        secondNextMedia:
          type:
            - object
            - 'null'
          description: 2つ後のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        previousMedia:
          type:
            - object
            - 'null'
          description: 1つ前のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        secondPreviousMedia:
          type:
            - object
            - 'null'
          description: 2つ前のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: '2026-01-01T00:00:00Z'
...
    MediaUpdateRequestDto:
      type: object
      description: メディア更新リクエスト
      properties:
        albumId:
          type:
            - integer
            - 'null'
          format: int64
          description: 所属アルバムID
        sharingGroupId:
          type:
            - integer
            - 'null'
          format: int64
          description: 共有範囲グループID
...
    MediaTagsUpdateRequestDto:
      type: object
      description: メディアのタグ一括更新リクエスト
      required:
        - tagIds
      properties:
        tagIds:
          type: array
          description: 紐付けるタグIDの一覧
          items:
            type: integer
            format: int64
          example:
            - 1
            - 2
            - 3
    TagRequestDto:
      type: object
      description: タグ作成/編集リクエスト
      required:
        - name
      properties:
        name:
          type: string
          description: タグ名
          minLength: 1
          maxLength: 100
          example: お出かけ
...
  responses:
    DefaultErrors:
      description: 予期しないエラーが発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Unauthorized:
      description: 認証エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    NotFound:
      description: 指定されたリソースが見つからない
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    InternalServerError:
      description: サーバー内部でエラー発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Forbidden:
      description: アクセス権限エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'